chatbot_api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
langchain-core==0.1.10
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.20.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
